+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -93,5 +93,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-26T16:26:20.368Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-27T09:45:48.496Z")
     r"""The date and time the resource was last updated"""