+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KB0FSWP3WYQ3NFYH46P1G72X"
+        "01KB2B96M8DQW2V4ARYS53KCK2"
     )
     r"""The unique identifier of the data source"""