+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KB0FSWP52QS53C7C2SA649RZ"
+        "01KB2B96MAS2C7N1720JEBCY5W"
     )
     r"""The unique identifier of the data source"""