+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KB0FSWPAFB1NB6DF4A1308EK"
+        "01KB2B96MEXNPYEBA1N0V9FFJ6"
     )
     r"""The unique identifier of the data source"""